edit: Add remaining missing nation change reporting
Print a message describing the actual change for keys 't' (nat_tgms)
and 's' (nat_status).  The message is necessary to give the deity a
chance to catch unexpected changes, e.g. a player reading telegrams
just before the deity edits nat_tgms.

Send a bulletin to the country for key 's'.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>