info/march: Correct and clarify
Don't claim the army stops when the leader stops.

Don't claim the lowest-numbered land unit is always the leader.

The prompt shows minimum mobility, not leader mobility.

Clarify that a destination sector is also accepted interactively, not
just on the command line.  Fix the example output, and update for
current code.

Replace incorrect landmine hit chance formula by a reference to "info
Hitchance".

Drop incorrect damage limit for missile interdiction, rely on the
reference to "info mission" instead.

The hit chance of missiles and pin-bombers interdicting land units is
not 100%, but depends on the marching land unit that is easiest to hit.

Clarify that interdiction damage is spread evenly among the marching
land units.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>